Kernel: SGR implementado corretamente no terminal

+++ kernel/include/vga.h
@@ -9,8 +9,8 @@
 #define VGA_WIDTH 80
 #define VGA_HEIGHT 25
 
-void vga_set_cursor(uint16_t x, uint16_t y);
-void vga_put_char(uint16_t x, uint16_t y, char c, uint8_t attributes);
-uint16_t vga_get_char(uint16_t x, uint16_t y);
+void vga_set_cursor(int x, int y);
+void vga_put_char(int x, int y, char c, uint8_t attributes);
+uint16_t vga_get_char(int x, int y);
 
 #endif /* VGA_H */
