+++ kernel/source/vga.c
@@ -9,7 +9,7 @@
 uint16_t *vga = (uint16_t *)0xB8000;
 
 /* Muda a posição do cursor no modo de texto VGA */
-void vga_set_cursor(uint16_t x, uint16_t y)
+void vga_set_cursor(int x, int y)
 {
 	uint16_t pos = y * VGA_WIDTH + x;
 
@@ -20,7 +20,7 @@ void vga_set_cursor(uint16_t x, uint16_t y)
 }
 
 /* Desenha um caractere diretamente no modo de texto VGA */
-void vga_put_char(uint16_t x, uint16_t y, char c, uint8_t attributes)
+void vga_put_char(int x, int y, char c, uint8_t attributes)
 {
 	if (x >= VGA_WIDTH || y >= VGA_HEIGHT)
 		return;
@@ -31,7 +31,7 @@ void vga_put_char(uint16_t x, uint16_t y, char c, uint8_t attributes)
 }
 
 /* Retorna um caractere de uma posição específica do modo de texto VGA */
-uint16_t vga_get_char(uint16_t x, uint16_t y)
+uint16_t vga_get_char(int x, int y)
 {
 	if (x >= VGA_WIDTH || y >= VGA_HEIGHT)
 		return 0;
